vendortraining/requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
djangorestframework>=3.10
django-mysql>=3.0
mysqlclient>=1.3.13
coreapi>=1.32.0
markdown>=3.0.0
pygments>=2.4.0
django-filter>=1.0.1
django-guardian>=1.1.1